requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
greenlet==1.1.2
idna==3.3
numpy==1.22.2
pandas==1.4.1
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
tzdata==2021.5
urllib3==1.26.8
